Extract `buildmodel.ts` to decouple unnecessary dependencies that we have in common.ts
