+++ src/compile/concat.ts
@@ -2,7 +2,7 @@ import {Config} from '../config';
 import {ConcatSpec, isVConcatSpec} from '../spec';
 import {VgLayout} from '../vega.schema';
 import {BaseConcatModel} from './baseconcat';
-import {buildModel} from './common';
+import {buildModel} from './buildmodel';
 import {parseConcatLayoutSize} from './layoutsize/parse';
 import {Model} from './model';
 import {RepeaterValue} from './repeater';
