+++ src/compile/repeat.ts
@@ -9,7 +9,7 @@ import {RepeatSpec} from '../spec';
 import {Dict, keys} from '../util';
 import {VgLayout} from '../vega.schema';
 import {BaseConcatModel} from './baseconcat';
-import {buildModel} from './common';
+import {buildModel} from './buildmodel';
 import {parseRepeatLayoutSize} from './layoutsize/parse';
 import {Model} from './model';
 import {RepeaterValue} from './repeater';
